+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/MergeAllOfTransformer.scala
@@ -1,0 +1,26 @@
+package edu.rit.cs.mmior.jsonoid.discovery
+
+import schemas._
+
+object MergeAllOfTransformer {
+  @SuppressWarnings(Array("org.wartremover.warts.Var"))
+  def transformSchema(
+      schema: JsonSchema[_]
+  )(implicit er: EquivalenceRelation): JsonSchema[_] = {
+    schema.transformProperties(
+      { case p @ ProductSchema(properties) =>
+        val schemaTypesProp = p.properties.get[ProductSchemaTypesProperty]
+        if (schemaTypesProp.all) {
+          var schema = schemaTypesProp.baseSchema
+          schemaTypesProp.schemaTypes.foreach { s =>
+            schema = schema.merge(s, Intersect)
+          }
+          schema
+        } else {
+          p
+        }
+      },
+      true
+    )
+  }
+}